riscv: Adding vdso data logic for placing extension information
Adding the logic for utilizing vdso data as the intermediate for
extension information display.

We are repurposing the HWCAP2 field as a pointer which points to an area
residing in the vdso data region, which contains a "sort-of" linked
list.

The caveats now is that we need to find a place for including the
vdso/data.h.

Signed-off-by: Ruinland Tsai <ruinland.tsai@sifive.com>